location done

+++ lib/providers/profile_provider.dart
@@ -200,14 +200,27 @@ class ProfileProvider extends StateNotifier<Profile?> {
     return friends.isNotEmpty;
   }
 
-  Future<void> createBlockedTime(profileId, sTime, eTime, title) async {
+  Future<void> createBlockedTime(profileId, sTime, eTime, title, String? eventId) async {
     final supabaseClient = (await supabase).client;
-    Map blockedTimeMap = {
+    Map blockedTimeMap;
+
+    if(eventId == null) {
+      blockedTimeMap = {
       'user_id': profileId,
       'start_time': sTime,
       'end_time': eTime,
       'block_title': title
-    };
+      };
+    } else {
+      blockedTimeMap = {
+      'user_id': profileId,
+      'start_time': sTime,
+      'end_time': eTime,
+      'block_title': title,
+      'event_id' : eventId
+      };
+    }
+    
     await supabaseClient.from('Availability').insert(blockedTimeMap);
   }
 
@@ -226,13 +239,19 @@ class ProfileProvider extends StateNotifier<Profile?> {
         .eq('availability_id', availId);
   }
 
-  Future<void> deleteBlockedTime(availId) async {
+  Future<void> deleteBlockedTime(String? availId, String? eventId) async {
     final supabaseClient = (await supabase).client;
-
-    await supabaseClient
-        .from('Availability')
-        .delete()
-        .eq('availability_id', availId);
+    if(eventId == null)
+      {await supabaseClient
+          .from('Availability')
+          .delete()
+          .eq('availability_id', availId!);
+      } else {
+        await supabaseClient
+          .from('Availability')
+          .delete()
+          .eq('event_id', eventId);
+      }
   }
 
   List<Availability> availavilityByMonth(int year, int month) {
